Eliminate useless lockquire() deco & other attempted synchronization stuff
The lockquire() decorator's twofold purpose was first to acquire an instance-communal class-variable asyncio.Lock() instance and apply it for the duration of a coroutine function's execution, and second to acquire a connection from the app's postgres connection pool and pass it for use by the function (here via the parameter I named 'conn', although it wasn't a keyword argument).

The former of these purposes was obsoleted twice: first when I found out that, due to asyncio.Lock()'s NOT being an R-Lock, calling a 'locked' function from within another would cause the app to hang indefinitely (necessitating a reboot), forcing me to set  on a considerable number of such functions to avoid this; and second when I recently found out that PostgreSQL BY DEFAULT locks write ops automatically to avoid race conditions (forcing operations to occur in sequence no matter how they are applied), reducing the effective need for this operation to absolutely nil.

The latter purpose was too obsoleted when I noticed, perusing repositories of better asyncio users than I, that for a simple single-command SQL query you will NEVER need to call pool.acquire() in order to utilize a connection object -- because the pool instance ALREADY has the basic execution and retrieval methods defined on it (which are equivalent in practice to acquiring a connection manually off of it and utilizing the same methods thereof). This meant that a large, large amount of multiline and multi-indent operations within my typedef.py class/instance methods could be noticeably reduced.

And with these two obsoletions, I no longer have any need for the deco -- and good riddance, I say! Hopefully this will let Python chill out with some of its 'got multiple values for argument 'self'' or 'No value provided for ...' exceptions.

Expect also a few more commits following this one to rectify the mistakes I invariably made here.